fix(checkout page): move file CheckoutPage.jsx to directory orders
